requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-dotenv==1.0.0
requests==2.31.0
stripe==7.0.0
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
